requirements.txt
antlr4-python3-runtime==4.9.3
einops==0.8.1
filelock==3.25.2
fsspec==2026.2.0
hydra-core==1.3.2
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
pandas==2.3.3
pillow==12.1.1
pyrosettacolabsetup==1.1.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.3
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.3
